Refactor character shooting and turret handling: rename CharacterShootingHandler and TurretsHandler, update turret rotation logic to use Float instead of Int, and streamline turret-related method calls for consistency.

+++ core/src/main/java/com/gadarts/returnfire/ecs/components/turret/TurretComponent.kt
@@ -16,7 +16,7 @@ class TurretComponent(
     var baseOffsetApplied = false
     private var currentShootingArm: Int = 1
     var turretRelativeRotation: Float = 0F
-    var turretRotating: Int = 0
+    var turretRotating: Float = 0F
 
 
     fun updateCurrentShootingArm(): Int {
